config: set HYPERV_NET as module
A race condition was showed when using the max supported synthetic NICs
on Hyper-V as Clear Linux kernel is too fast and the HW is not ready.

Set hyperv network driver as module helps HW to be ready.
This change is set only meanwhile found a better solution.

Note: Max supported synthetic NICs is 8.

Signed-off-by: Miguel Bernal Marin <miguel.bernal.marin@linux.intel.com>